[aggregator] Do layout matching in change_set only sporadically (#11204)
* Randomly check layout matches in change_set.rs

* Addressing PR comments

* Invoking randomly_check_layout_matches function in executor.rs
