+++ aptos-move/aptos-vm-types/Cargo.toml
@@ -24,6 +24,7 @@ claims = { workspace = true }
 either = { workspace = true }
 move-binary-format = { workspace = true }
 move-core-types = { workspace = true }
+rand = { workspace = true }
 serde = { workspace = true }
 
 [dev-dependencies]
